refactor: move rule helpers to src/rules/utils/
src/rules/ had 14 files mixing 4 user-facing rule factories with
3 internal helpers (match-paths, string-or-regex-matches,
trim-history) plus all their tests. Group the helpers under utils/
so the directory listing distinguishes "rules" from "rule helpers."
Pure file move + import path updates.
