+++ src/rules/utils/match-paths.test.ts
