+++ src/rules/utils/string-or-regex-matches.test.ts
