+++ src/rules/utils/string-or-regex-matches.ts
